Move exposed calculator functions to root package
